make it configurable which request are handled
Adds two new options:
* extensions (array of strings)
  if extension of requested file matches a value in array file is handled by node-phpcgi
* includePath (regular expression pattern)
  if path matches includePath request is handled even if extension is not in extensions

Tests for all cases are added. Not sure if the test that assets are not handled (or more
general that files neither included by `extensions` nor `includePath` are not handled)
is okay like that. It waits 1,5s if request ist handled if not it jumps in and says:
"Everything fine." Not sure if it's the best way to test for a timeout. Found these
approach here:
http://stackoverflow.com/questions/17647069/test-for-expected-timeout-in-mocha/17648268#17648268